more cleanup of errors, combine test_error_wrapper.py and test_error.py
